+++ NEWS.md
@@ -16,7 +16,7 @@
 
 ## Bug fixes
 
-- `standardize()` for models does not standardize offset predictors or responses when offsets are detected ( #396 ).
+- `standardize()` for models with offsets standardizes offset variables according to `include_response` and `two_sd` ( #396 ).
 - `eta_squared()`: fixed a bug that caused `afex_aov` models with more than 2 within-subject factors to return incorrect effect sizes for the lower level factors ( #389 ).
 
 # effectsize 0.5.0